Restructure Layer (#41)
* Add Components and Connections to Layer

Include validation and test updates. Remove missed Layer reference from
Component.

* Remove unnecessary functions from Validation

Slow ID uniqueness verification and map validation were useful only
before the restructuring of the data model.

* Remove 'valid' test from Parsing Tests

The structure of Layer has changed, so what used to be a valid Layer
(only a name and ID) is not anymore. This will be added back in after
the Parsing is updated.

Close #37
Close #34
Close #40

* Add JSDoc and @babel/preset-env modules

After installing JSDoc as a dependency, babel/preset-env seemed to be
missing.

* Make no Connections/Components on Layers valid

A Layer can exist without any Connections or Components, so it is a
warning rather than invalid.