Add Basic SVG generation (#84)
* Clean up webpage CSS

Multiple styles where only one was needed.

* Install Paper.js to repo

Close #74

* Add auto Architecture sizing to ParchmintParser

Currently the validation schema does not allow us to add the size of
the architecture, so we will set the size according to the farthest
components/connections.

* Add print function to Layer

Each Layer will have its own PaperScope object to print separate SVGs.

* Implement Component Feature printing

Currently the Component Feature is represented by a black bounding box.

* Implement Connection Segment drawing

Connection Segments are represented by black lines, the width of which
is defined by the Parchmint.

* Add name to missing feature/segment warnings

Components and Connections previously warned about not having features,
but did not specify which they were.

* Fix ParchmintParser Connection Segment error

Connections were looking for their segments with their Layer as the key,
but the Connection ID is the key.

* Install jest-canvas-mock package

Need to mock the canvas HTML object to run tests.

* Update parse button functionality on web page

Pushing parse now generates the SVGs after a successful Parchmint parse.

* Remove Travis docs test

Currently fails, and needs to be fixed. Will be fixed by issue #72.

* Change temp var value to segment in Connection

* Change console log to warn in validation functions

Suppressed console.warn in associated tests.

* Create Config module

Config module gives us an easy way to access and modify the global SVG
drawing variables. It should be simple to set these from either the
webpage or by loading a config.json.

Update Connection/Component Features with references to the module.

* Add device bounding box to output SVG

* Add input fields for device height and width

User should be able to specify the height and width of the device if
they so choose. This will most likely be removed when device x- and
y-spans are added to the Parchmint standard.

* Implement functionality for height/width inputs

If the given inputs are less than the values determined by the Parchmint
Parser, the device size is set to the determined values.

* Add labels for each SVG image generated.